requirements.txt
boto3==1.17.12
botocore==1.20.12
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
future==0.18.2
idna==2.10
jmespath==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.2
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
pytz==2021.1
regex==2020.11.13
requests==2.25.1
s3transfer==0.3.4
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
threadpoolctl==2.1.0
tqdm==4.65.0
typing-extensions==3.7.4.3
urllib3==1.26.5
transformers==4.29.2
pandas==1.1.5
fitlog==0.9.13